refactor(core): group phase classes under compiler/ subfolder
Moves Parse, Transform, Generate (and their tests) into
packages/core/src/compiler/, matching the storages/ pattern (subfolder,
no index.ts, callers import each file directly). Internal-only move — no
public surface change.

Adds a patch changeset.
